# Projet_Examen/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portail Scolaire API",
        "description": "Documentation de l'API du portail scolaire",
        "contact": {
            "email": "diabydaj@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/absences": {
            "get": {
                "tags": [
                    "Absences"
                ],
                "summary": "Liste des absences",
                "operationId": "b54d45f0d5b3eb84fcbc42c72bed408d",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Absences"
                ],
                "summary": "Créer une nouvelle absence",
                "operationId": "643ad2101a5afd13087ef6ed2c5ff1cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "eleve_id",
                                    "date",
                                    "motif"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Absence créée"
                    }
                }
            }
        },
        "/api/absences/{id}": {
            "get": {
                "tags": [
                    "Absences"
                ],
                "summary": "Afficher une absence spécifique",
                "operationId": "a243d6238afc2df1ce661ad38132c083",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Absence non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Absences"
                ],
                "summary": "Mettre à jour une absence",
                "operationId": "bfc6576ddf8a06c862bbef471d8b4da0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "motif"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Absence mise à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Absences"
                ],
                "summary": "Supprimer une absence",
                "operationId": "25e317727636714a0612684ebf5319b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Absence supprimée"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}